Fixed issue where some values of 9 minutes after the hour was not formatting properly

+++ src/org/redout/solunarlib/StringUtil.java
@@ -14,7 +14,7 @@ public class StringUtil {
 	        d = 0.0;
 	    }
 	/*convert times to a string*/
-	    if (d < 9.5) {
+	    if (d.intValue() < 9.5) {
 	        return(i.intValue() + ":" + "0" +d.intValue());
 	    }
 	    else {
